# specs/001-channel-membership-verification/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: VPass - Channel Membership Card Issuance API (MVP)
  description: |
    Member-facing endpoints for issuing digitally signed YouTube membership cards.

    Organizer verification, automated revocation, and multi-platform support will be described in follow-up specifications.
  version: 0.1.0
servers:
  - url: https://vpass.example.com
    description: Production (placeholder)
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /cards/{card_id}:
    get:
      tags: [Cards]
      summary: Retrieve a single membership card
      parameters:
        - in: path
          name: card_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Card details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipCard'
        '401':
          description: No active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Card not found or does not belong to the member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cards/{card_id}/qr:
    get:
      tags: [Cards]
      summary: Retrieve QR code for a membership card
      parameters:
        - in: path
          name: card_id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: format
          schema:
            type: string
            enum: [svg, png, data_url]
          description: Preferred QR output format (default svg)
      responses:
        '200':
          description: QR code payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QRCodeResponse'
        '401':
          description: No active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Card not found or does not belong to the member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    MembershipCard:
      type: object
      required:
        - id
        - issuer_id
        - member_id
        - membership_level_label
        - membership_confirmed_at
        - verification_comment_id
        - verification_video_id
        - issued_at
      properties:
        id:
          type: string
          format: uuid
        issuer_id:
          type: string
          format: uuid
        member_id:
          type: string
          format: uuid
        membership_level_label:
          type: string
        membership_confirmed_at:
          type: string
          format: date-time
        verification_comment_id:
          type: string
          description: Membership verification marker (historically a YouTube comment ID)
        verification_video_id:
          type: string
        issued_at:
          type: string
          format: date-time
        qr_signature:
          type: string
          description: Hex-encoded HMAC signature
        qr_payload:
          type: object
          additionalProperties: true
        supplementary_data:
          type: object
          additionalProperties: true
          nullable: true
...
    QRCodeResponse:
      type: object
      properties:
        format:
          type: string
          enum: [svg, png, data_url]
        value:
          type: string
          description: Base64 or inline SVG string depending on format
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true